+++ server/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Fixed
 
-- references: consider listing a section in parts as a reference.
+- references: consider listing a section in `${buildout:parts}` as a reference.
 
 ## [0.2.0] - 2020-02-12